# src/tests/EshopGuard.Api.Tests/Snapshots/openapi-v1.json
{
  "components": {
    "schemas": {
...
      "ChangeRoleRequest": {
        "properties": {
          "role": {
            "type": [
              "null",
              "string"
            ]
          }
        },
        "required": [
          "role"
        ],
        "type": "object"
      },
...
      "CreateInvitationRequest": {
        "properties": {
          "email": {
            "type": [
              "null",
              "string"
            ]
          },
          "locale": {
            "type": [
              "null",
              "string"
            ]
          },
          "role": {
            "type": [
              "null",
              "string"
            ]
          }
        },
        "required": [
          "email",
          "role",
          "locale"
        ],
        "type": "object"
      },
      "CreateTenantRequest": {
        "properties": {
          "market": {
            "type": [
              "null",
              "string"
            ]
          },
          "name": {
            "type": [
              "null",
              "string"
            ]
          }
        },
        "required": [
          "name",
          "market"
        ],
        "type": "object"
      },
...
      "InvitationDto": {
        "properties": {
          "acceptedAt": {
            "format": "date-time",
            "type": [
              "null",
              "string"
            ]
          },
          "email": {
            "type": "string"
          },
          "expiresAt": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "invitedBy": {
            "format": "uuid",
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "role",
          "invitedBy",
          "expiresAt",
          "acceptedAt",
          "status"
        ],
        "type": "object"
      },
...
      "MemberDto": {
        "properties": {
          "displayName": {
            "type": [
              "null",
              "string"
            ]
          },
          "email": {
            "type": "string"
          },
          "invitedBy": {
            "format": "uuid",
            "type": [
              "null",
              "string"
            ]
          },
          "joinedAt": {
            "format": "date-time",
            "type": "string"
          },
          "lastLoginAt": {
            "format": "date-time",
            "type": [
              "null",
              "string"
            ]
          },
          "role": {
            "type": "string"
          },
          "userId": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "userId",
          "email",
          "displayName",
          "role",
          "invitedBy",
          "joinedAt",
          "lastLoginAt"
        ],
        "type": "object"
      },
...
      "RenameTenantRequest": {
        "properties": {
          "name": {
            "type": [
              "null",
              "string"
            ]
          },
          "version": {
            "format": "uint32",
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ]
          }
        },
        "required": [
          "name",
          "version"
        ],
        "type": "object"
      },
...
      "TenantDto": {
        "properties": {
          "countryCode": {
            "type": "string"
          },
          "currency": {
            "type": [
              "null",
              "string"
            ]
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "legalName": {
            "type": [
              "null",
              "string"
            ]
          },
          "locale": {
            "type": "string"
          },
          "marketCode": {
            "type": "string"
          },
          "myRole": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "version": {
            "format": "uint32",
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ]
          }
        },
        "required": [
          "id",
          "name",
          "legalName",
          "countryCode",
          "marketCode",
          "locale",
          "currency",
          "status",
          "myRole",
          "version"
        ],
        "type": "object"
      },
...
      "TransferOwnershipRequest": {
        "properties": {
          "userId": {
            "format": "uuid",
            "type": [
              "null",
              "string"
            ]
          }
        },
        "required": [
          "userId"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "EshopGuard.Api | v1",
    "version": "1.0.0"
  },
  "openapi": "3.1.1",
  "paths": {
...
    "/api/t/{tenantId}": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "tenant"
        ],
        "x-problem-codes": [
          "auth.unauthenticated",
          "csrf.invalid",
          "internal_error",
          "rate_limited",
          "request.https_required",
          "request.invalid",
          "tenant.not_found",
          "tenant.suspended"
        ]
      },
      "patch": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/RenameTenantRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "tenant"
        ],
        "x-problem-codes": [
          "auth.forbidden_role",
          "auth.unauthenticated",
          "concurrency.conflict",
          "csrf.invalid",
          "internal_error",
          "rate_limited",
          "request.https_required",
          "request.invalid",
          "tenant.not_found",
          "tenant.suspended",
          "validation.failed"
        ]
      }
    },
    "/api/t/{tenantId}/invitations": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/InvitationDto"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "tenant"
        ],
        "x-problem-codes": [
          "auth.forbidden_role",
          "auth.unauthenticated",
          "csrf.invalid",
          "internal_error",
          "rate_limited",
          "request.https_required",
          "request.invalid",
          "tenant.not_found",
          "tenant.suspended"
        ]
      },
      "post": {
        "parameters": [
          {
            "in": "path",
            "name": "tenantId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/CreateInvitationRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitationDto"
                }
              }
            },
            "description": "Created"
          }
        },
        "tags": [
          "tenant"
        ],
        "x-problem-codes": [
          "auth.forbidden_role",
          "auth.unauthenticated",
          "csrf.invalid",
          "email.send_failed",
          "internal_error",
          "invitation.already_member",
          "locale.not_enabled",
          "membership.role_not_allowed",
          "rate_limited",
          "request.https_required",
          "request.invalid",
          "tenant.not_found",
          "tenant.suspended",
          "validation.failed"
        ]
      }
    },
    "/api/t/{tenantId}/invitations/{invitationId}": {
      "delete": {
        "parameters": [
          {
            "in": "path",
            "name": "invitationId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "tags": [
          "tenant"
        ],
        "x-problem-codes": [
          "auth.forbidden_role",
          "auth.unauthenticated",
          "csrf.invalid",
          "internal_error",
          "invitation.not_found",
          "invitation.used",
          "rate_limited",
          "request.https_required",
          "request.invalid",
          "tenant.not_found",
          "tenant.suspended"
        ]
      }
    },
    "/api/t/{tenantId}/invitations/{invitationId}/resend": {
      "post": {
        "parameters": [
          {
            "in": "path",
            "name": "invitationId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "tags": [
          "tenant"
        ],
        "x-problem-codes": [
          "auth.forbidden_role",
          "auth.unauthenticated",
          "csrf.invalid",
          "email.send_failed",
          "internal_error",
          "invitation.not_found",
          "invitation.used",
          "rate_limited",
          "request.https_required",
          "request.invalid",
          "tenant.not_found",
          "tenant.suspended"
        ]
      }
    },
...
    "/api/t/{tenantId}/members/{userId}": {
      "delete": {
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "tags": [
          "tenant"
        ],
        "x-problem-codes": [
          "auth.unauthenticated",
          "csrf.invalid",
          "internal_error",
          "membership.last_owner",
          "membership.not_found",
          "membership.role_not_allowed",
          "rate_limited",
          "request.https_required",
          "request.invalid",
          "tenant.not_found",
          "tenant.suspended"
        ]
      },
      "patch": {
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/ChangeRoleRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "tenant"
        ],
        "x-problem-codes": [
          "auth.forbidden_role",
          "auth.unauthenticated",
          "csrf.invalid",
          "internal_error",
          "membership.last_owner",
          "membership.not_found",
          "membership.role_not_allowed",
          "rate_limited",
          "request.https_required",
          "request.invalid",
          "tenant.not_found",
          "tenant.suspended",
          "validation.failed"
        ]
      }
    },
    "/api/t/{tenantId}/ownership-transfer": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/TransferOwnershipRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "tags": [
          "tenant"
        ],
        "x-problem-codes": [
          "auth.forbidden_role",
          "auth.unauthenticated",
          "csrf.invalid",
          "internal_error",
          "membership.not_found",
          "rate_limited",
          "request.https_required",
          "request.invalid",
          "tenant.not_found",
          "tenant.suspended",
          "validation.failed"
        ]
      }
    },
    "/api/tenants": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/CreateTenantRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantDto"
                }
              }
            },
            "description": "Created"
          }
        },
        "tags": [
          "tenants"
        ],
        "x-problem-codes": [
          "auth.unauthenticated",
          "csrf.invalid",
          "internal_error",
          "market.unknown",
          "rate_limited",
          "request.https_required",
          "request.invalid",
          "tenant.limit_reached",
          "validation.failed"
        ]
      }
    }
  },
  "servers": [
    {
      "url": "https://localhost/"
    }
  ],
...
}